[newjersey/doula-pm#398] Test that pdf cover page has actual text
Previously, we had a bug where the cover page of the output pdf could not be accessibly read - it seemed to be an image, that was not text highlight-able or readable by a screenreader. The bug itself has since be been fixed. However, this is bug is easy to not notice and annoying to fill the form and manually check each time.

This commit thus adds a cypress test that tests that tests that the downloaded form has a text-readable cover page.

This commit resolves newjersey/doula-pm#398.
