+++ package.json
@@ -12,11 +12,7 @@
     "prettier:check": "prettier . --check",
     "prepare": "husky",
     "cypress:run": "cypress run",
-    "cypress:run:productionFlags": "cypress run --env productionFlags=true",
-    "cypress:run:websiteUnavailable": "cypress run --env websiteUnavailable=true",
     "cypress:gui": "cypress open",
-    "cypress:gui:productionFlags": "cypress open --env productionFlags=true",
-    "cypress:gui:websiteUnavailable": "cypress open --env websiteUnavailable=true",
     "test": "env $(cat cypress/.env.test-production-flags | xargs) vitest --run",
     "typecheck": "tsc --build",
     "configureVpc": "npx tsx scripts/configureVpc.ts",
@@ -61,6 +57,7 @@
     "jiti": "2.6.1",
     "jsdom": "27.4.0",
     "lint-staged": "latest",
+    "pdfjs-dist": "5.4.449",
     "prettier": "latest",
     "prettier-plugin-jsdoc": "1.8.0",
     "prettier-plugin-organize-imports": "latest",
